test: update tests and place them in packages
